Deduplicate restricted visibility comment spec helpers
